chore: wow build works
