+++ package.json
@@ -99,7 +99,7 @@
   "bundlesize": [
     {
       "path": "packages/algoliasearch/dist/algoliasearch.umd.js",
-      "maxSize": "8.4KB"
+      "maxSize": "9.1KB"
     },
     {
       "path": "packages/algoliasearch/dist/algoliasearch-lite.umd.js",
