+++ packages/algoliasearch/src/types/AlgoliaSearchOptions.ts
@@ -3,8 +3,6 @@ import { ClientTransporterOptions } from '@algolia/client-common';
 import { PersonalizationClientOptions } from '@algolia/client-personalization';
 import { SearchClientOptions } from '@algolia/client-search';
 
-import { TransformationOptions } from '.';
-
 type Credentials = { readonly appId: string; readonly apiKey: string };
 export type WithoutCredentials<TClientOptions extends Credentials> = Omit<
   TClientOptions,
@@ -17,8 +15,7 @@ export type OptionalCredentials<TClientOptions extends Credentials> = Omit<
   Pick<Partial<TClientOptions>, keyof Credentials>;
 
 export type AlgoliaSearchOptions = Partial<ClientTransporterOptions> &
-  WithoutCredentials<SearchClientOptions> &
-  TransformationOptions;
+  WithoutCredentials<SearchClientOptions>;
 
 export type InitAnalyticsOptions = Partial<ClientTransporterOptions> &
   OptionalCredentials<AnalyticsClientOptions>;
