+++ packages/algoliasearch/src/types/index.ts
@@ -3,4 +3,3 @@
  */
 
 export * from './AlgoliaSearchOptions';
-export * from './Ingestion';
